+++ R/Defaults.R
@@ -439,11 +439,8 @@ sosDefaultColorPalette <- c("#A6CEE3", "#1F78B4", "#B2DF8A", "#33A02C",
 
 sosDefaultReferenceFrameSensorDescription <- "urn:ogc:def:crs:EPSG:4326"
 
-# default values for non-standard stuff
-sosDefaultGetBindingParamLatest <- NA_character_ # e.g. time=latest
-
 #
-#
+# Defaults ----
 #
 SosDefaults <- function() {
   .defaults <- list(sosDefaultCharacterEncoding,
@@ -461,8 +458,7 @@ SosDefaults <- function() {
                     sosDefaultColumnNameLat,
                     sosDefaultColumnNameLon,
                     sosDefaultColumnNameSRS,
-                    sosDefaultReferenceFrameSensorDescription,
-                    sosDefaultGetBindingParamLatest)
+                    sosDefaultReferenceFrameSensorDescription)
   names(.defaults) <- list("sosDefaultCharacterEncoding",
                            "sosDefaultDescribeSensorOutputFormat",
                            "sosDefaultGetCapSections",
@@ -478,8 +474,7 @@ SosDefaults <- function() {
                            "sosDefaultColumnNameLat",
                            "sosDefaultColumnNameLon",
                            "sosDefaultColumnNameSRS",
-                           "sosDefaultReferenceFrameSensorDescription",
-                           "sosDefaultGetBindingParamLatest")
+                           "sosDefaultReferenceFrameSensorDescription")
 
   return(.defaults)
 }